Update Solver.java
I wish I knew how to code better. There are sooooooo many if statements. I was planning on making two more methods like colPossibilities, but it would add another 300 lines. I am also submitting this just to save it, but I need to go back and delete a whole bunch of stuff because it is just redundant.